refactor: domain-name remaining-ops-9; peel heartbeat prompts
Rename task-store/remaining-ops-9 to task-commit-associations (U3 first
domain migration). Extract agent-heartbeat system prompts/procedures into
agent-heartbeat-prompts.ts with stable re-exports (U6 slice).
